Split site.py into a focused site package.
Decompose build, markdown rendering, HTML shell, and backlinks into site/ submodules while preserving public imports from wiki.site.

Closes #54
