+++ src/wiki/site/__init__.py
@@ -1,0 +1,51 @@
+"""Site-building logic for compiling raw Markdown wikis into HTML virtual structures."""
+
+from ..schemas.site import InfoboxRow, TocItem, VirtualPage, WikiSite
+from .build import build_site
+from .html import (
+    DEFAULT_THEME_COLOR,
+    build_index_html,
+    build_infobox_rows,
+    build_page_html,
+    build_web_manifest,
+    resolved_site_theme_color,
+    serialize_web_manifest,
+)
+from .markdown import (
+    DEFAULT_MINIMAL_PAGE_LAYOUT,
+    INLINE_CSS,
+    extract_title,
+    humanize_route,
+    render_copyable_pre,
+    render_outline_title,
+    render_wiki_markdown,
+    split_by_headings,
+    strip_leading_title_heading,
+)
+
+# Test-only / internal exports kept for compatibility
+from .html import _build_logo_svg  # noqa: F401
+
+__all__ = [
+    "DEFAULT_MINIMAL_PAGE_LAYOUT",
+    "DEFAULT_THEME_COLOR",
+    "INLINE_CSS",
+    "InfoboxRow",
+    "TocItem",
+    "VirtualPage",
+    "WikiSite",
+    "build_index_html",
+    "build_infobox_rows",
+    "build_page_html",
+    "build_site",
+    "build_web_manifest",
+    "extract_title",
+    "humanize_route",
+    "render_copyable_pre",
+    "render_outline_title",
+    "render_wiki_markdown",
+    "resolved_site_theme_color",
+    "serialize_web_manifest",
+    "split_by_headings",
+    "strip_leading_title_heading",
+]
